Windows Batch File Alternative
This is to allow for this to be useful for both windows and linux.

What you would need to do, is to modify the Linux Makefile to read off
config.txt instead of being hardcoded into the makefile.

This will make this academic boilerplate more useful for collab between
academics who uses windows or linux